Add files via upload
It is a small project after lesson 'Working with multiple DataFrames', to create a Funnel for page visitors from a T-shirt company. Its function is to check how many percentage of Page visitors will move to the next levels of the Funnel from Visit to Cart, from Cart to Checkout, and from Checkout to Purchase, giving more meaning on visitors behavior, and percentage of pottential customers. 
It also gives average time for customers giving their decisions on buying T-shirt.
It's quite fun to work with, inspecting DataFrames, and see how it can be applied in real business.